add index file, add package documentation and remove api-extractor warnings
